requirements.txt
fastapi
uvicorn
google-generativeai
yahoo_fin
requests
langchain
